Bug 547328 - JPA NamedStoredProcedure call getOutputParameterValue class cast exception (#451) (#453)
(cherry picked from commit 45cee6cd8733655c1ca31cffaf7127628f9c6937)

+++ foundation/org.eclipse.persistence.core/src/org/eclipse/persistence/internal/databaseaccess/DatabasePlatform.java
@@ -1319,7 +1319,7 @@ public class DatabasePlatform extends DatasourcePlatform {
         } else if (javaType == ClassConstants.INTEGER) {
             return Types.INTEGER;
         } else if (javaType == ClassConstants.LONG) {
-            return Types.INTEGER;
+            return Types.BIGINT;
         } else if (javaType == ClassConstants.NUMBER) {
             return Types.DECIMAL;
         } else if (javaType == ClassConstants.SHORT ) {
